Improve dependency declarations for grpc-plugins
For packages that require a grpc plugin for protoc for code generation,
use RpmWith in their PkgSpec.py files to define the dependency on the
installed name of the required plugin rather than depending directly on
the rpm name. Should the plugins move to a different rpm in the future,
as they did somewhere between grpc versions 1.21 and 1.41, this will
help avoid breakage. Since the generated protobuf files are the ones
that depend on the grpc plugins, attach the plugin dependencies to them,
not to the libraries they're compiled into.

Fix the regular expressions in gnmi PkgSpec.py used to identify the
protoc-generated files to allow both pb and pb2, since files of both
forms are generated in gnmi.

In grpc.spec, list the plugins individually, rather than using a wildcard,
so pkgdeps can find them.

Remove all c-ares rpms from /src/External, as they are obsolete now that
/src/c-ares provides newer versions. The newer version in /src/c-ares was
added recently for the recent grpc upgrade to 1.41.1.

http://reviewboard/r/355325
integrate all member packages from stevev.grpc.10@27167104 (launched at change 27164077)
eos-trunk@27175847 autobuild 14844287

Change-Id: I1eea5a2a683aa96b5f7bd121f1e1901df6e0b73d