Fixed an issue where multiple MacsecProfile RPC fields were writing to the same underlying SDK data members and overwritting each other.
In MacsecProfile the deprecated fields bypass_lldp and allow_unprotected write to the same underlying SDK attributes as lldp_bypass_level and traffic_policy respectively. This means that in the types conversion whichever is converted last will control the configured value in the SDK. This is an issue as gRPC cannot distinguish when a field is set with the default value or when it is unset. Therefore, configured fields can be overwritten by the default values in unconfigured fields. The fix is to stop supporting the deprecated fields completely. This ensure that there is a one to one mapping between the SDK and RPC fields and that user configuration can not be overwritten.

The deprecated fields are marked as reserved in the proto file so that the server will ignore them and any clients compiling or recompiling can not use them. However, existing clients will not break.

Change-Id: Ia82cfbf289f9baeb67bfb7783ad4bf914711280e