Adding dump for Intf RPC Service.
This MUT introduces the ability to dump the current Intf manager state in EosSdkRpc when first subscribing to the watcher for events. Clients can use this to build up a local copy of the state of the Intf manager which they can then use the watch events to update.

The WatchIntfRequest and WatchIntfResponse proto messages have been updated. The WatchIntfRequest now allows the user to specify which fields they want in the dump. While the WatchIntfResponse has been reworked slightly to make supporting the dump easier. No protobuf compatibility breaking changes have been made.

Updated IntfService to use the adapter pattern for the request handlers.

Change-Id: Ifa1547a1cf86a18e00c56652803e660e794bc3c4
Reviewed-on: https://gitarband-gerrit.infra.corp.arista.io/c/EosSdkRpcProtos/+/887625
Reviewed-by: Anonymous Arastra <arastra@arista.com>
Reviewed-by: Philip Bradish <bradishp@arista.com>