Fixed issues with the default values of MplsLabel in EosSdkRpc.
MplsRouteLabel should have a default value of 1048576 instead of 0. This was previously missed when adding default parameter values to the EosSdkRpc that match those defined in EosSdk as it is specified in mpls.cpp instead of types/mpls_impl.h.

Also removed an erroneous default value for the MplsRouteKey label list. This was given a default value consisting of a single MplsLabel with value 0 instead of a MplsLabel list of length 0.

Change-Id: Id6e250e9a2a7106585ca7d7a7cc66f66e4320839
Reviewed-on: https://gitarband-gerrit.infra.corp.arista.io/c/EosSdkRpcProtos/+/948356
Reviewed-by: Philip Bradish <bradishp@arista.com>